requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==1.4.46
asyncpg==0.27.0
alembic==1.10.4